+++ R/globalvariables.R
@@ -5,4 +5,5 @@
 #' @name globalVariables
 #' @title Global Variables Declaration
 #' @keywords internal
-utils::globalVariables(c("motif", "start", "ID", "length", "seq_order", "end", "name", "class", "class.x", "support", "x", "y", "yend", "Group"))
+utils::globalVariables(c("motif", "start", "ID", "length", "seq_order", "end", "name", 
+                         "class", "class.x", "support", "x", "y", "yend", "Group","frequency"))